pyproject.toml
[tool.poetry]
name = "english_learning_app"
version = "0.1.0"
description = ""
authors = ["irisuinwl <iris.dimension.eee@gmail.com>"]
readme = "README.md"
packages = [{include = "english_learning_app"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.21.0"
PyYAML = "^6.0"
langchain = "^0.0.142"
google-cloud-texttospeech = "^2.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
